Enable XDP fragment support
Introduced in Linux 5.18, this comes with support for jumbo frame that crosses multiple memory pages, but limited direct packet access (first memory page size of packet data is directly accessible). Since we only do that on the headers, and restoring data is done using bpf_xdp_{load,store}_bytes, it should be safe to enable this for better compatibility.